requirements-dev.txt
bs4
colorama
coverage
flake8
gitpython
mypy
PyGithub
pytest
pytest-ordering
requests
